requirements.txt
fastapi>=0.115.0
pydantic>=2.8.0
python-docx>=1.1.0
openpyxl>=3.1.0
PyYAML>=6.0.0
typer>=0.12.0
pytest>=8.0.0
